Fix Unit Tests
