+++ API/DTOs/SeriesDetail/UserReviewDto.cs
@@ -58,5 +58,4 @@ public class UserReviewDto
     /// If this review is External, which Provider did it come from
     /// </summary>
     public ScrobbleProvider Provider { get; set; } = ScrobbleProvider.Kavita;
-    public ReviewAuthority Authority { get; set; } = ReviewAuthority.User;
 }
